Extract dag node logic to shared lib
