+++ airflow-core/src/airflow/_shared/definitions
@@ -1,0 +1,1 @@
+../../../../shared/definitions/src/airflow_shared/definitions
