+++ task-sdk/src/airflow/sdk/_shared/definitions
@@ -1,0 +1,1 @@
+../../../../../shared/definitions/src/airflow_shared/definitions
